# apps/people-service/openapi/people-api.yaml
openapi: 3.0.3
info:
  title: People REST API
  version: 1.0.0
  description: |
    CRUD API for Person records (first name, last name, age).
    All endpoints require a Keycloak Bearer token with the `people-crud` role.
servers:
  - url: https://people-backend-${WORKSHOP_NAMESPACE}.${CLUSTER_ROUTER_BASE}
    description: OpenShift route (workshop namespace)
  - url: http://localhost:8080
    description: Local Quarkus dev mode
security:
  - bearerAuth: []
paths:
  /api/people:
    get:
      operationId: listPeople
      summary: List all people
      tags: [People]
      responses:
        "200":
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createPerson
      summary: Create a person
      tags: [People]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonRequest"
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/people/{id}:
    parameters:
      - $ref: "#/components/parameters/PersonId"
    get:
      operationId: getPerson
      summary: Get a person by ID
      tags: [People]
      responses:
        "200":
          description: Person found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updatePerson
      summary: Update a person
      tags: [People]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonRequest"
      responses:
        "200":
          description: Person updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePerson
      summary: Delete a person
      tags: [People]
      responses:
        "204":
          description: Person deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PersonId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    PersonRequest:
      type: object
      required: [firstName, lastName, age]
      properties:
        firstName:
          type: string
          example: Ada
        lastName:
          type: string
          example: Lovelace
        age:
          type: integer
          minimum: 0
          maximum: 150
          example: 36
    Person:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstName:
          type: string
          example: Ada
        lastName:
          type: string
          example: Lovelace
        age:
          type: integer
          example: 36
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
    Forbidden:
      description: Token valid but missing people-crud role
    NotFound:
      description: Person not found
    BadRequest:
      description: Invalid request payload